pyproject.toml
[project]
name = "api-test-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "allure-pytest>=2.15.0",
    "allure-python-commons>=2.15.0",
    "colorlog>=6.9.0",
    "httpx>=0.28.1",
    "jsonpath-python>=1.0.6",
    "jsonschema>=4.25.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pyyaml>=6.0.2",
]
